feat: db 데이터를 es에 색인하기 위한 logstash 설정 추가

+++ src/main/java/kr/kro/airbob/search/dto/AccommodationSearchRequest.java
@@ -68,7 +68,7 @@ public class AccommodationSearchRequest {
 		}
 
 		// 총 인원 수 계산 (유아, 펫 제외)
-		public int getTotalGuests() {
+		public double getTotalGuests() {
 			return (adultOccupancy != null ? adultOccupancy : 0) +
 				(childOccupancy != null ? childOccupancy : 0);
 		}
